requirements.txt
flask
torch
torchvision
pillow
numpy
opencv-python-headless
pyyaml
matplotlib
requests
scipy
tqdm
seaborn
